openfl-develop/requirements.txt
absl-py==0.13.0
aiofiles==0.5.0
aniso8601==7.0.0
argon2-cffi==21.1.0
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distro==1.4.0
dnspython==2.1.0
docker==5.0.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.68.1
filelock==3.3.1
flatten-json==0.1.13
gast==0.3.3
gdown==3.13.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.30.0
grpcio-tools==1.30.0
h11==0.12.0
httplib2==0.14.0
httptools==0.1.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.3.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==2.11.3
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.4.0
more-itertools==4.2.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.3
oauthlib==3.1.0
opt-einsum==3.3.0
orjson==3.6.3
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.2.4
pluggy==1.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyftpdlib==1.5.6
Pygments==2.10.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==9.1.0
rsa==4.7.2
Rx==1.6.1
scikit-learn==0.24.2
scipy==1.7.1
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
testresources==2.0.1
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
Twisted==18.9.0
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.25.8
uvicorn==0.13.4
uvloop==0.16.0
wadllib==1.3.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.13.2
zipp==1.0.0
zope.interface==4.7.1
